requirements.txt
asgiref==3.7.2
blessed==1.19.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
colorama==0.4.4
cryptography==41.0.3
Django==4.2.4
django-rest-knox==4.2.0
django-split-settings==1.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
fire==0.4.0
halo==0.0.31
idna==3.2
log-symbols==0.0.14
pathspec==0.9.0
phonenumbers==8.13.18
Pillow==10.0.0
pycparser==2.21
Pygments==2.10.0
PyJWT==1.7.1
python-dotenv==1.0.0
python-editor==1.0.4
pytz==2021.3
readchar==2.0.1
regex==2023.6.3
requests==2.26.0
requests-toolbelt==0.9.1
runflare==1.1.5
six==1.16.0
spinners==0.0.24
sqlparse==0.4.4
termcolor==1.1.0
typing_extensions==4.7.1
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.7
wcwidth==0.2.5
websockets==10.0

pymongo~=4.4.1